+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- Issue a notification and mark experiments as unavailable when they
+  are selected in the GUI but the code is unable to read data from them
 ### Fixed
 - Issue that could cause the code to leave orphan processes behind
   after cancelling plots/multiPlots, which would lead to memory leaks